+++ apps/web-vite/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web-vite
 
+## 0.0.11
+
+### Patch Changes
+
+- Updated dependencies
+  - @hbwow/hooks@2.1.0
+
 ## 0.0.10
 
 ### Patch Changes