+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hbwow/hooks
 
+## 2.1.0
+
+### Minor Changes
+
+- useSearchAndTable添加storeKey，用于全局缓存搜索条件，有则全局缓存，无则不需要缓存 (通常使用路由值，防止重复)
+
 ## 2.0.6
 
 ### Patch Changes